UI and partial shortcuts
When using tab, the button's text in the installer gets underlined.
Program should no longer crash when attempting installation without administrator privileges.
Button in install turns red when there are errors.

"Exit" menu option added under "File" in the MainWindow.
Separators added in menus in MainWindow.
Shortcut text (only text) for menu options added in MainWindow.
When pressing Alt and a key for a menu that is in the MainWindow's menubar, there are keys that can be pressed to trigger a click (instead of using the mouse). (Example: press and hold "Alt" then press "F", then release both and press "E")

Shortcuts are mostly implemented (all structs and most WINAPI functions), only error is that interfaces can't be pointed to.
Some more research may be needed for an interface that can be pointed to, or a DLL that gets programmed in C++ could also do the trick.
Path for Start Menu shortcut might not be correct.